+++ monitoring_hub/customers/__init__.py
@@ -1,0 +1,1 @@
+"""Customer configuration helpers."""